Add device flow
